+++ KeeAnywhere/OAuth2/OAuth2Form.cs
@@ -90,7 +90,7 @@ namespace KeeAnywhere.OAuth2
 
         public Exception LastException { get; set; }
 
-        private async void OnNavigating(object sender, WebBrowserNavigatingEventArgs e)
+        private void OnNavigating(object sender, WebBrowserNavigatingEventArgs e)
         {
             Debug.WriteLine("Navigating " + e.Url);
         }